+++ src/main/java/jamdoggie/betterbattletowers/entity/MobGolem.java
@@ -22,7 +22,7 @@ import java.util.List;
 
 import static jamdoggie.betterbattletowers.entity.damage.DamageInstance.inst;
 import static jamdoggie.betterbattletowers.util.MathUtil.posGausssianIntBounded;
-import static jamdoggie.betterbattletowers.worldgen.util.LootTable.LAPIZ;
+import static jamdoggie.betterbattletowers.config.LootTable.LAPIZ;
 import static net.minecraft.core.Global.TICKS_PER_SECOND;
 
 @SuppressWarnings("java:S2160")
